speedup win32 link checks
Fixed linux coverage
