+++ .coveragerc
@@ -22,6 +22,7 @@ exclude_lines =
 
 omit =
     ubelt/tests/*
+    ubelt/_win32_links.py
     ubelt/_internal/dynamic_make_init.py
     ubelt/_internal/static_autogen.py
     ubelt/_internal/__init__.py
