+++ appveyor.yml
@@ -9,7 +9,7 @@ environment:
 
     - PYTHON: "C:\\Python35"
     - PYTHON: "C:\\Python27"
-    - PYTHON: "C:\\Python36"
+    #- PYTHON: "C:\\Python36"
     #- PYTHON: "C:\\Python33"
     #- PYTHON: "C:\\Python33-x64"
     #  DISTUTILS_USE_SDK: "1"
