+++ ubelt/util_platform.py
@@ -5,7 +5,6 @@ from os.path import expanduser
 from os.path import join
 from os.path import normpath
 from os.path import islink
-from os.path import isdir
 import os
 import sys
 import pipes
@@ -21,6 +20,9 @@ WIN32  = sys.platform.startswith('win32')
 LINUX  = sys.platform.startswith('linux')
 DARWIN = sys.platform.startswith('darwin')
 POSIX = 'posix' in sys.builtin_module_names
+
+if WIN32:
+    from ubelt import _win32_links
 
 
 def platform_resource_dir():
@@ -315,10 +317,10 @@ def symlink(real_path, link_path, overwrite=False, verbose=0):
             # However, if it is a non-junction directory then do not overwrite
             if verbose:
                 print('link location already exists')
-            is_junc = _win32_is_junction(link)
+            is_junc = _win32_links._win32_is_junction(link)
             if os.path.isdir(link):
                 if is_junc:
-                    pointed = _win32_read_junction(link)
+                    pointed = _win32_links._win32_read_junction(link)
                     if path == pointed:
                         if verbose:
                             print('...and is a junction that points to the same place')
@@ -336,7 +338,7 @@ def symlink(real_path, link_path, overwrite=False, verbose=0):
 
             elif os.path.isfile(link):
                 # if os.stat(link).st_ino == os.stat(path).st_ino:
-                if _win32_hardlinks_equal(link, path):
+                if _win32_links._win32_hardlinks_equal(link, path):
                     if verbose:
                         print('...and is a hard link that points to the same place')
                     return link
@@ -353,47 +355,10 @@ def symlink(real_path, link_path, overwrite=False, verbose=0):
                 if exists(link):
                     raise IOError('Link already exists')
 
-        _win32_symlink2(path, link)
+        _win32_links._win32_symlink2(path, link, verbose=verbose)
     else:
         os.symlink(path, link)
     return link
-
-
-def _win32_hardlinks_equal(fpath1, fpath2):  # nocover
-    """
-    http://timgolden.me.uk/python/win32_how_do_i/see_if_two_files_are_the_same_file.html
-    """
-    import win32file
-
-    def get_read_handle(filename):
-        if os.path.isdir(filename):
-            dwFlagsAndAttributes = win32file.FILE_FLAG_BACKUP_SEMANTICS
-        else:
-            dwFlagsAndAttributes = 0
-        return win32file.CreateFile(filename, win32file.GENERIC_READ,
-                                    win32file.FILE_SHARE_READ, None,
-                                    win32file.OPEN_EXISTING,
-                                    dwFlagsAndAttributes, None)
-
-    def get_unique_id(hFile):
-        (attributes,
-         created_at, accessed_at, written_at,
-         volume,
-         file_hi, file_lo,
-         n_links,
-         index_hi, index_lo) = win32file.GetFileInformationByHandle(hFile)
-        return volume, index_hi, index_lo
-
-    try:
-        hFile1 = get_read_handle(fpath1)
-        hFile2 = get_read_handle(fpath2)
-        are_equal = (get_unique_id(hFile1) == get_unique_id(hFile2))
-    except Exception:
-        raise
-    finally:
-        hFile1.Close()
-        hFile2.Close()
-    return are_equal
 
 
 def _readlink(link):
@@ -403,9 +368,6 @@ def _readlink(link):
         # On modern operating systems, we should never get here. (I think)
         warnings.warn('Reading symlinks seems to not be supported')
         raise
-
-
-__win32_can_symlink__ = None
 
 
 def _can_symlink(verbose=0):  # nocover
@@ -422,296 +384,9 @@ def _can_symlink(verbose=0):  # nocover
         >>> print(_can_symlink(verbose=1))
     """
     if WIN32:
-        return False
-        return _win32_can_symlink(verbose)
+        return _win32_links._win32_can_symlink(verbose)
     else:
         return True
-
-
-def _win32_can_symlink(verbose=0):  # nocover
-    global __win32_can_symlink__
-    if verbose:
-        print('__win32_can_symlink__ = {!r}'.format(__win32_can_symlink__))
-    if __win32_can_symlink__ is not None:
-        return __win32_can_symlink__
-
-    tempdir = ensure_app_cache_dir('ubelt', '_win32_can_symlink')
-    # import shutil
-    # shutil.rmtree(tempdir)
-    util_io.delete(tempdir)
-    util_path.ensuredir(tempdir)
-
-    dpath = join(tempdir, 'dpath')
-    fpath = join(tempdir, 'fpath.txt')
-    dlink = join(tempdir, 'dlink')
-    flink = join(tempdir, 'flink.txt')
-
-    util_path.ensuredir(dpath)
-    util_io.touch(fpath)
-
-    try:
-        _win32_symlink(dpath, dlink)
-        can_symlink_directories = True
-        # os.path.islink(dlink)
-    except OSError:
-        can_symlink_directories = False
-    if verbose:
-        print('can_symlink_directories = {!r}'.format(can_symlink_directories))
-
-    try:
-        _win32_symlink(fpath, flink)
-        can_symlink_files = True
-        # os.path.islink(flink)
-    except OSError:
-        can_symlink_files = False
-    if verbose:
-        print('can_symlink_files = {!r}'.format(can_symlink_files))
-
-    assert int(can_symlink_directories) + int(can_symlink_files) != 1, (
-        'can do one but not both. Unexpected {} {}'.format(
-            can_symlink_directories, can_symlink_files))
-
-    try:
-        # test that we can create junctions, even if symlinks are disabled
-        _win32_junction(dpath, join(tempdir, 'djunc'))
-        _win32_junction(fpath, join(tempdir, 'fjunc.txt'))
-    except Exception:
-        warnings.warn('We cannot create junctions either!')
-        raise
-
-    # Cleanup the test directory
-    util_io.delete(tempdir)
-
-    can_symlink = can_symlink_directories and can_symlink_files
-    __win32_can_symlink__ = can_symlink
-    if not can_symlink:
-        warnings.warn('Cannot make real symlink. Falling back to junction')
-
-    if verbose:
-        print('can_symlink = {!r}'.format(can_symlink))
-        print('__win32_can_symlink__ = {!r}'.format(__win32_can_symlink__))
-    return can_symlink
-
-
-def _win32_junction(path, link):
-    """
-    On older (pre 10) versions of windows we need admin privledges to make
-    symlinks, however junctions seem to work.
-
-    For paths we do a junction (softlink) and for files we use a hard link
-
-    TODO:
-        # Need code to test if path is a junction
-        # https://stackoverflow.com/questions/18883892/batch-file-windows-cmd-exe-test-if-a-directory-is-a-link-symlink
-
-        # TODO: need to test readlink for junctions
-    """
-    from ubelt import util_cmd
-    # if platform.release() == '7':
-    if isdir(path):
-        # try using a junction (soft link)
-        command = 'mklink /J "{}" "{}"'.format(link, path)
-    else:
-        # try using a hard link
-        command = 'mklink /H "{}" "{}"'.format(link, path)
-    info = util_cmd.cmd(command, shell=True)
-    if info['ret'] != 0:
-        from ubelt import util_format
-        print('Failed command:')
-        print(info['command'])
-        print(util_format.repr2(info, nl=1))
-        raise OSError(str(info))
-
-
-def _win32_is_junction(path):
-    """
-    Determines if a path is a win32 junction
-    """
-    if not exists(path):
-        return False
-    try:
-        return _win32_read_junction(path) is not None
-    except (ValueError, OSError):
-        return False
-
-
-def _win32_read_junction(path):
-    """
-    Returns the location that the junction points, raises ValueError if path is
-    not a junction.
-
-    CommandLine:
-        python -m ubelt.util_platform _win32_read_junction
-
-    Example:
-        >>> # DISABLE_DOCTEST
-        >>> # xdoc: +REQUIRES(WIN32)
-        >>> import ubelt as ub
-        >>> root = ub.ensure_app_cache_dir('ubelt', 'win32_symlink')
-        >>> ub.delete(root)
-        >>> ub.ensuredir(root)
-        >>> fpath = join(root, 'fpath.txt')
-        >>> dpath = join(root, 'dpath')
-        >>> fjunc = join(root, 'fjunc.txt')
-        >>> djunc = join(root, 'djunc')
-        >>> ub.touch(fpath)
-        >>> ub.ensuredir(dpath)
-        >>> ub.ensuredir(join(root, 'djunc_fake'))
-        >>> ub.ensuredir(join(root, 'djunc_fake with space'))
-        >>> ub.touch(join(root, 'djunc_fake with space file'))
-        >>> _win32_junction(fpath, fjunc)
-        >>> _win32_junction(dpath, djunc)
-        >>> # thank god colons are not allowed
-        >>> djunc2 = join(root, 'djunc2 [with pathological attrs]')
-        >>> _win32_junction(dpath, djunc2)
-        >>> _win32_is_junction(djunc)
-        >>> ub.writeto(join(djunc, 'afile.txt'), 'foo')
-        >>> assert ub.readfrom(join(dpath, 'afile.txt')) == 'foo'
-        >>> ub.writeto(fjunc, 'foo')
-
-
-        # >>> if not _win32_can_symlink(verbose=1):
-        # ...     import pytest
-        # ...     pytest.skip()
-
-    """
-    if not WIN32:
-        return False
-    if not exists(path):
-        if six.PY2:
-            raise OSError('Cannot find path={}'.format(path))
-        else:
-            raise FileNotFoundError('Cannot find path={}'.format(path))
-    target_name = os.path.basename(path)
-    for type_or_size, name, pointed in _win32_dir(path, '*'):
-        if type_or_size == '<JUNCTION>' and name == target_name:
-            return pointed
-    raise ValueError('not a junction')
-
-
-def _win32_rmtree(path, verbose=0):
-    """
-    rmtree for win32 that treats junctions like directory symlinks.
-    The junction removal portion may not be safe on race conditions.
-
-    There is a known issue that prevents shutil.rmtree from
-    deleting directories with junctions.
-    https://bugs.python.org/issue31226
-    """
-    def _rmjunctions(root):
-        subdirs = []
-        for type_or_size, name, pointed in _win32_dir(root):
-            if type_or_size == '<DIR>':
-                subdirs.append(name)
-            elif type_or_size == '<JUNCTION>':
-                # remove any junctions as we encounter them
-                # os.unlink(join(root, name))
-                os.rmdir(join(root, name))
-        # recurse in all real directories
-        for name in subdirs:
-            _rmjunctions(join(root, name))
-
-    if _win32_is_junction(path):
-        if verbose:
-            print('Deleting <JUNCTION> directory="{}"'.format(path))
-        os.rmdir(path)
-    else:
-        if verbose:
-            print('Deleting directory="{}"'.format(path))
-        # first remove all junctions
-        _rmjunctions(path)
-        # now we can rmtree as normal
-        import shutil
-        shutil.rmtree(path)
-
-
-def _win32_dir(path, star=''):
-    from ubelt import util_cmd
-    import re
-    wrapper = 'cmd /S /C "{}"'  # the /S will preserve all inner quotes
-    command = 'dir /-C "{}"{}'.format(path, star)
-    wrapped = wrapper.format(command)
-    info = util_cmd.cmd(wrapped, shell=True)
-    if info['ret'] != 0:
-        from ubelt import util_format
-        print('Failed command:')
-        print(info['command'])
-        print(util_format.repr2(info, nl=1))
-        raise OSError(str(info))
-    # parse the output of dir to get some info
-    # Remove header and footer
-    lines = info['out'].split('\n')[5:-3]
-    splitter = re.compile('( +)')
-    for line in lines:
-        parts = splitter.split(line)
-        date, sep, time, sep, ampm, sep, type_or_size, sep = parts[:8]
-        name = ''.join(parts[8:])
-        # if type is a junction then name will also contain the linked loc
-        if name == '.' or name == '..':
-            continue
-        if type_or_size in ['<JUNCTION>', '<SYMLINKD>', '<SYMLINK>']:
-            # colons cannot be in path names, so use that to find where
-            # the name ends
-            pos = name.find(':')
-            bpos = name[:pos].rfind('[')
-            name = name[:bpos - 1]
-            pointed = name[bpos + 1:-1]
-            yield type_or_size, name, pointed
-        else:
-            yield type_or_size, name, None
-
-
-def _win32_symlink(path, link):  # nocover
-    from ubelt import util_cmd
-    # # This strategy requires admin permissions
-    # import ctypes
-    # csl = ctypes.windll.kernel32.CreateSymbolicLinkW
-    # csl.argtypes = (ctypes.c_wchar_p, ctypes.c_wchar_p, ctypes.c_uint32)
-    # csl.restype = ctypes.c_ubyte
-    # flags = 1 if isdir(path) else 0
-    # if csl(link, path, flags) == 0:
-    #     raise ctypes.WinError('cannot create win32 symlink')
-
-    # This strategy requires development mode to be on I believe.
-    if isdir(path):
-        # directory symbolic link
-        command = 'mklink /D "{}" "{}"'.format(link, path)
-    else:
-        # file symbolic link
-        command = 'mklink "{}" "{}"'.format(link, path)
-    info = util_cmd.cmd(command, shell=True)
-    if info['ret'] != 0:
-        from ubelt import util_format
-        permission_msg = 'You do not have sufficient privledges'
-        if permission_msg not in info['err']:
-            print('Failed command:')
-            print(info['command'])
-            print(util_format.repr2(info, nl=1))
-        raise OSError(str(info))
-
-
-def _win32_symlink2(path, link, allow_fallback=True):  # nocover
-    """
-    Perform a real symbolic link if possible. However, on most versions of
-    windows you need special privledges to create a real symlink. Therefore, we
-    try to create a symlink, but if that fails we fallback to using a junction.
-
-    AFAIK, the main difference between symlinks and junctions are that symlinks
-    can reference relative or absolute paths, where as junctions always
-    reference absolute paths. Not 100% on this though. Windows is weird.
-
-    Note that junctions will not register as links via `islink`, but I
-    believe real symlinks will.
-
-    References:
-        https://stackoverflow.com/questions/6260149/os-symlink-support-in-windows
-        https://msdn.microsoft.com/en-us/library/windows/desktop/aa365006(v=vs.85).aspx
-        https://superuser.com/a/902082/215232
-    """
-    if _can_symlink():
-        _win32_symlink(path, link)
-    else:
-        _win32_junction(path, link)
 
 
 if __name__ == '__main__':
